Remove capturing lambdas
