+++ core/optaplanner-core-impl/src/main/java/org/optaplanner/core/impl/domain/solution/cloner/DeepCloningFieldCloner.java
@@ -1,20 +1,20 @@
 package org.optaplanner.core.impl.domain.solution.cloner;
 
 import java.lang.reflect.Field;
-import java.util.function.Consumer;
 
 final class DeepCloningFieldCloner implements FieldCloner {
 
     static final FieldCloner INSTANCE = new DeepCloningFieldCloner();
 
     @Override
-    public <C> void clone(DeepCloningUtils deepCloningUtils, Field field, Class<? extends C> instanceClass, C original, C clone,
-            Consumer<Object> deferredValueConsumer) {
+    public <C> Unprocessed clone(DeepCloningUtils deepCloningUtils, Field field, Class<? extends C> instanceClass, C original,
+            C clone) {
         Object originalValue = FieldCloner.getFieldValue(original, field);
         if (isDeepCloneField(deepCloningUtils, field, instanceClass, originalValue)) { // Deffer filling in the field.
-            deferredValueConsumer.accept(originalValue);
+            return new Unprocessed(clone, field, originalValue);
         } else { // Shallow copy.
             FieldCloner.setFieldValue(clone, field, originalValue);
+            return null;
         }
     }
 
@@ -24,11 +24,6 @@ final class DeepCloningFieldCloner implements FieldCloner {
             return false;
         }
         return deepCloningUtils.getDeepCloneDecision(field, fieldInstanceClass, originalValue.getClass());
-    }
-
-    @Override
-    public boolean mayDeferClone() {
-        return true;
     }
 
     private DeepCloningFieldCloner() {
