+++ core/optaplanner-core-impl/src/main/java/org/optaplanner/core/impl/domain/solution/cloner/Unprocessed.java
@@ -1,0 +1,17 @@
+package org.optaplanner.core.impl.domain.solution.cloner;
+
+import java.lang.reflect.Field;
+
+final class Unprocessed {
+
+    final Object bean;
+    final Field field;
+    final Object originalValue;
+
+    public Unprocessed(Object bean, Field field, Object originalValue) {
+        this.bean = bean;
+        this.field = field;
+        this.originalValue = originalValue;
+    }
+
+}
